pyproject.toml
[project]
name = "pytorch-computer-vision"
version = "0.1.0"
description = "PyTorch-based fruit and vegetable classification project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=1.10.0",
    "torchvision>=0.11.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
]

[project.scripts]
predict = "src.predict:main"
gui-predict = "src.gui_predict:main"
